Safari-609.1.11.1/LayoutTests/crypto/subtle/rsa-import-pkcs8-key-export-pkcs8-key-expected.txt

Test importing a PKCS8 RSA-OAEP key and then export it in PKCS8 format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(privateKey) is pkcs8KeyHex
PASS successfullyParsed is true

TEST COMPLETE
